gofmt all the things!

+++ replican/fs/util.go
@@ -1,4 +1,3 @@
-
 package fs
 
 import (
@@ -22,33 +21,39 @@ func Move(src string, dst string) (err os.Error) {
 	if _, err = os.Stat(dst); err == nil {
 		os.Remove(dst)
 	}
-	
+
 	if err = os.Rename(src, dst); err != nil {
 		linkErr, isLinkErr := err.(*os.LinkError)
-		if !isLinkErr { return err }
-		
-		if causeErr, isErrno := linkErr.Error.(os.Errno); 
-				isErrno && causeErr == syscall.EXDEV {
-			srcF, err := os.Open(src)
-			if err != nil { return err }
-			defer srcF.Close()
-			
-			dstF, err := os.Create(dst)
-			if err != nil { return err }
-			defer dstF.Close()
-			
-			_, err = io.Copy(dstF, srcF)
-			if err != nil { return err }
-			
-			srcF.Close()
-			err = os.Remove(src)
-			
+		if !isLinkErr {
 			return err
 		}
-		
+
+		if causeErr, isErrno := linkErr.Error.(os.Errno); isErrno && causeErr == syscall.EXDEV {
+			srcF, err := os.Open(src)
+			if err != nil {
+				return err
+			}
+			defer srcF.Close()
+
+			dstF, err := os.Create(dst)
+			if err != nil {
+				return err
+			}
+			defer dstF.Close()
+
+			_, err = io.Copy(dstF, srcF)
+			if err != nil {
+				return err
+			}
+
+			srcF.Close()
+			err = os.Remove(src)
+
+			return err
+		}
+
 		return err
 	}
-	
+
 	return nil
 }
-
